console: Avoid repetition
