+++ pkg/webui/console/containers/sidebar/navigation/app-list-side-navigation.js
@@ -12,46 +12,20 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-import React, { useCallback, useContext, useState } from 'react'
+import React, { useContext } from 'react'
 import { useSelector } from 'react-redux'
-
-import { IconPlus } from '@ttn-lw/components/icon'
-import SectionLabel from '@ttn-lw/components/sidebar/section-label'
-import SideNavigation from '@ttn-lw/components/sidebar/side-menu'
-import Button from '@ttn-lw/components/button'
-
-import RequireRequest from '@ttn-lw/lib/components/require-request'
-
-import sharedMessages from '@ttn-lw/lib/shared-messages'
-import PropTypes from '@ttn-lw/lib/prop-types'
-import useBookmark from '@ttn-lw/lib/hooks/use-bookmark'
-
-import { getAllBookmarks } from '@console/store/actions/user-preferences'
 
 import { selectApplicationBookmarks } from '@console/store/selectors/user-preferences'
 import { selectUserId } from '@console/store/selectors/logout'
 
 import SidebarContext from '../context'
 
-const Bookmark = ({ bookmark }) => {
-  const { title, ids, path, icon } = useBookmark(bookmark)
-
-  return <SideNavigation.Item title={title === '' ? ids.id : title} path={path} icon={icon} />
-}
-
-Bookmark.propTypes = {
-  bookmark: PropTypes.shape({}).isRequired,
-}
+import TopEntitiesSection from './top-entities-section'
 
 const AppListSideNavigation = () => {
-  const [showMore, setShowMore] = useState(false)
   const topEntities = useSelector(state => selectApplicationBookmarks(state))
   const { isMinimized } = useContext(SidebarContext)
   const userId = useSelector(selectUserId)
-
-  const handleShowMore = useCallback(async () => {
-    setShowMore(showMore => !showMore)
-  }, [])
 
   if (isMinimized || topEntities.length === 0) {
     // Rendering an empty div to prevent the shadow of the search bar
@@ -60,28 +34,7 @@ const AppListSideNavigation = () => {
     return <div />
   }
 
-  return (
-    <RequireRequest requestAction={getAllBookmarks(userId)}>
-      <SideNavigation>
-        <SectionLabel label={sharedMessages.topApplications} icon={IconPlus} onClick={() => null} />
-        {topEntities.slice(0, 6).map(bookmark => (
-          <Bookmark key={bookmark.created_at} bookmark={bookmark} />
-        ))}
-        {showMore &&
-          topEntities.length > 6 &&
-          topEntities
-            .slice(6, topEntities.length)
-            .map(bookmark => <Bookmark key={bookmark.created_at} bookmark={bookmark} />)}
-        {topEntities.length > 6 && (
-          <Button
-            message={showMore ? sharedMessages.showLess : sharedMessages.showMore}
-            onClick={handleShowMore}
-            className="c-text-neutral-light ml-cs-xs mt-cs-xs fs-s"
-          />
-        )}
-      </SideNavigation>
-    </RequireRequest>
-  )
+  return <TopEntitiesSection topEntities={topEntities} userId={userId} />
 }
 
 export default AppListSideNavigation
