+++ pkg/webui/console/containers/sidebar/navigation/general-side-navigation.js
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-import React, { useCallback, useContext, useState } from 'react'
+import React, { useContext } from 'react'
 import { useSelector } from 'react-redux'
 
 import { PAGE_SIZES } from '@ttn-lw/constants/page-sizes'
@@ -22,18 +22,11 @@ import {
   IconLayoutDashboard,
   IconUserShield,
   IconKey,
-  IconPlus,
   IconInbox,
 } from '@ttn-lw/components/icon'
 import SideNavigation from '@ttn-lw/components/sidebar/side-menu'
-import SectionLabel from '@ttn-lw/components/sidebar/section-label'
-import Button from '@ttn-lw/components/button'
-
-import RequireRequest from '@ttn-lw/lib/components/require-request'
 
 import sharedMessages from '@ttn-lw/lib/shared-messages'
-import PropTypes from '@ttn-lw/lib/prop-types'
-import useBookmark from '@ttn-lw/lib/hooks/use-bookmark'
 
 import {
   checkFromState,
@@ -42,26 +35,15 @@ import {
 } from '@console/lib/feature-checks'
 import getCookie from '@console/lib/table-utils'
 
-import { getAllBookmarks } from '@console/store/actions/user-preferences'
-
 import { selectUser, selectUserIsAdmin } from '@console/store/selectors/logout'
 import { selectBookmarksList } from '@console/store/selectors/user-preferences'
 
 import SidebarContext from '../context'
 
-const Bookmark = ({ bookmark }) => {
-  const { title, ids, path, icon } = useBookmark(bookmark)
-
-  return <SideNavigation.Item title={title === '' ? ids.id : title} path={path} icon={icon} />
-}
-
-Bookmark.propTypes = {
-  bookmark: PropTypes.shape({}).isRequired,
-}
+import TopEntitiesSection from './top-entities-section'
 
 const GeneralSideNavigation = () => {
   const { isMinimized } = useContext(SidebarContext)
-  const [showMore, setShowMore] = useState(false)
   const topEntities = useSelector(state => selectBookmarksList(state))
   const isUserAdmin = useSelector(selectUserIsAdmin)
   const user = useSelector(selectUser)
@@ -71,10 +53,6 @@ const GeneralSideNavigation = () => {
   const mayHandleApiKeys = useSelector(state =>
     user ? checkFromState(mayViewOrEditApiKeys, state) : false,
   )
-
-  const handleShowMore = useCallback(async () => {
-    setShowMore(showMore => !showMore)
-  }, [])
 
   const orgPageSize = getCookie('organizations-list-page-size')
   const orgParam = `?page-size=${orgPageSize ? orgPageSize : PAGE_SIZES.REGULAR}`
@@ -118,26 +96,7 @@ const GeneralSideNavigation = () => {
         )}
       </SideNavigation>
       {!isMinimized && topEntities.length > 0 && (
-        <RequireRequest requestAction={getAllBookmarks(user.ids.user_id)}>
-          <SideNavigation className="mt-cs-xs">
-            <SectionLabel label={sharedMessages.topEntities} icon={IconPlus} onClick={() => null} />
-            {topEntities.slice(0, 6).map((bookmark, index) => (
-              <Bookmark key={index} bookmark={bookmark} />
-            ))}
-            {showMore &&
-              topEntities.length > 6 &&
-              topEntities
-                .slice(6, topEntities.length)
-                .map(bookmark => <Bookmark key={bookmark.created_at} bookmark={bookmark} />)}
-            {topEntities.length > 6 && (
-              <Button
-                message={showMore ? sharedMessages.showLess : sharedMessages.showMore}
-                onClick={handleShowMore}
-                className="c-text-neutral-light ml-cs-xs mt-cs-xs fs-s"
-              />
-            )}
-          </SideNavigation>
-        </RequireRequest>
+        <TopEntitiesSection topEntities={topEntities} userId={user.ids.user_id} />
       )}
     </>
   )
