+++ pkg/webui/console/containers/sidebar/navigation/gtw-list-side-navigation.js
@@ -12,26 +12,20 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-import React, { useCallback, useContext, useState } from 'react'
+import React, { useContext } from 'react'
 import { useSelector } from 'react-redux'
 
-import { IconPlus } from '@ttn-lw/components/icon'
-import SectionLabel from '@ttn-lw/components/sidebar/section-label'
 import SideNavigation from '@ttn-lw/components/sidebar/side-menu'
-import Button from '@ttn-lw/components/button'
 
-import RequireRequest from '@ttn-lw/lib/components/require-request'
-
-import sharedMessages from '@ttn-lw/lib/shared-messages'
 import useBookmark from '@ttn-lw/lib/hooks/use-bookmark'
 import PropTypes from '@ttn-lw/lib/prop-types'
-
-import { getAllBookmarks } from '@console/store/actions/user-preferences'
 
 import { selectGatewayBookmarks } from '@console/store/selectors/user-preferences'
 import { selectUserId } from '@console/store/selectors/logout'
 
 import SidebarContext from '../context'
+
+import TopEntitiesSection from './top-entities-section'
 
 const Bookmark = ({ bookmark }) => {
   const { title, ids, path, icon } = useBookmark(bookmark)
@@ -44,41 +38,15 @@ Bookmark.propTypes = {
 }
 
 const GtwListSideNavigation = () => {
-  const [showMore, setShowMore] = useState(false)
   const topEntities = useSelector(state => selectGatewayBookmarks(state))
   const { isMinimized } = useContext(SidebarContext)
   const userId = useSelector(selectUserId)
-
-  const handleShowMore = useCallback(async () => {
-    setShowMore(showMore => !showMore)
-  }, [])
 
   if (isMinimized || topEntities.length === 0) {
     return <div />
   }
 
-  return (
-    <RequireRequest requestAction={getAllBookmarks(userId)}>
-      <SideNavigation>
-        <SectionLabel label={sharedMessages.topGateways} icon={IconPlus} onClick={() => null} />
-        {topEntities.slice(0, 6).map(bookmark => (
-          <Bookmark key={bookmark.created_at} bookmark={bookmark} />
-        ))}
-        {showMore &&
-          topEntities.length > 6 &&
-          topEntities
-            .slice(6, topEntities.length)
-            .map(bookmark => <Bookmark key={bookmark.created_at} bookmark={bookmark} />)}
-        {topEntities.length > 6 && (
-          <Button
-            message={showMore ? sharedMessages.showLess : sharedMessages.showMore}
-            onClick={handleShowMore}
-            className="c-text-neutral-light ml-cs-xs mt-cs-xs fs-s"
-          />
-        )}
-      </SideNavigation>
-    </RequireRequest>
-  )
+  return <TopEntitiesSection topEntities={topEntities} userId={userId} />
 }
 
 export default GtwListSideNavigation
