+++ pkg/webui/console/containers/top-entities-dashboard-panel/top-entities-panel.styl
@@ -40,3 +40,7 @@
       margin-bottom: 56px
   &-cell
     width: 33%
+    &-extended
+      width: 100%
+    &-small
+      width: 12%
